requirements.txt
numpy==2.1.3
Pillow==11.0.0
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.14.1
timm==1.0.11
torch==2.0.1
torchvision==0.15.2
